Redux test :mask:
