--- src/components/Visual.tsx
@@ -1,21 +1,0 @@
-/** @jsx jsx */
-import { jsx, css } from "@emotion/core"
-
-const Visual = () => {
-  return (
-    <div
-      css={css`
-        display: flex;
-        align-items: center;
-        font-size: 3rem;
-        justify-content: center;
-        background-color: #364156;
-        color: #d66853;
-      `}
-    >
-      VISUAL
-    </div>
-  )
-}
-
-export default Visual
