--- src/components/elems/HeaderWrapper.tsx
@@ -1,12 +1,0 @@
-import styled from "@emotion/styled"
-
-const HeaderWrapper = styled.div`
-  background-color: #11151c;
-  color: #fff;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 0 8px 0 3px;
-`
-
-export default HeaderWrapper
