+++ src/components/graph/GraphButton.tsx
@@ -1,0 +1,13 @@
+import styled from "@emotion/styled"
+
+const GraphButton = styled.button`
+  position: absolute;
+  right: 10px;
+  top: 10px;
+  width: 150px;
+  zindex: 4;
+  background-color: #212d40;
+  color: #fff;
+`
+
+export default GraphButton
