+++ src/components/header/Save.tsx
@@ -1,0 +1,7 @@
+/** @jsx jsx */
+import { jsx } from "@emotion/core"
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
+
+const Save = () => <FontAwesomeIcon icon={["fal", "save"]} size="lg" />
+
+export default Save
